Moved GlobalClock into the System namespace.
