+++ src/mindtouch.bench/Program.cs
@@ -24,7 +24,6 @@ using System.Collections.Generic;
 using System.Text;
 using System.IO;
 using System.Diagnostics;
-using MindTouch.Threading.Timer;
 
 namespace MindTouch.Tools {
     class DreamBench {
