+++ src/mindtouch.blueprints/Program.cs
@@ -27,7 +27,6 @@ using System.Text.RegularExpressions;
 
 using MindTouch.Dream;
 using MindTouch.Tasking;
-using MindTouch.Threading.Timer;
 using MindTouch.Web;
 using MindTouch.Xml;
