+++ src/mindtouch.dream/Collections/ExpiringDictionary.cs
@@ -24,7 +24,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using MindTouch.Tasking;
-using MindTouch.Threading.Timer;
 
 namespace MindTouch.Collections {
     /// <summary>
