+++ src/mindtouch.dream/Tasking/AsyncUtil.cs
@@ -29,7 +29,6 @@ using System.Threading;
 using MindTouch.Dream;
 using MindTouch.IO;
 using MindTouch.Threading;
-using MindTouch.Threading.Timer;
 
 namespace MindTouch.Tasking {
     using Yield = IEnumerator<IYield>;
