+++ src/mindtouch.dream/Tasking/TaskTimerFactory.cs
@@ -23,7 +23,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using MindTouch.Dream;
-using MindTouch.Threading.Timer;
 
 namespace MindTouch.Tasking {
