+++ src/mindtouch.dream/dream/servicebase.cs
@@ -28,7 +28,6 @@ using System.Text;
 
 using MindTouch.Security.Cryptography;
 using MindTouch.Tasking;
-using MindTouch.Threading.Timer;
 using MindTouch.Web;
 using MindTouch.Xml;
