+++ src/tests/DreamMisc/BlockingQueueTests.cs
@@ -23,7 +23,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using MindTouch.Collections;
-using MindTouch.Threading.Timer;
 using NUnit.Framework;
 using System.Linq;
