+++ src/tests/DreamMisc/ExpiringDictionaryTests.cs
@@ -23,7 +23,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using MindTouch.Threading.Timer;
 using log4net;
 using MindTouch.Collections;
 using MindTouch.Extensions.Time;
